mostrar comentarios de forma descendiente, cuando se agrega un comentario, se actualiza el promedioRating del producto, index muestra productos por createdAt descendiente (consigna) y por rating promedio (extra), search result por descripcion

+++ Prog2-G3/database/models/Comment.js
@@ -17,7 +17,7 @@ module.exports = function (Sequelize,DataTypes){ //el modelo exporta una funcion
         },
         rating:{
             notNull: true,
-            type: DataTypes.DECIMAL(2,1),
+            type: DataTypes.FLOAT,
         },
         FkPhoneId: {
             notNull: true,
